// desktop/src/open-api/model/recognitionTask.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecognitionTaskStage } from './recognitionTaskStage';
import { RecognitionTaskStatus } from './recognitionTaskStatus';


export interface RecognitionTask { 
    id: number;
    clientRequestId: string;
    fileName: string;
    fileSize: number;
    groupId: number;
    ownerUserId: number;
    version: number;
    status: RecognitionTaskStatus;
    stage: RecognitionTaskStage;
    createdAt: string;
    updatedAt: string;
    queuedAt?: string;
    startedAt?: string;
    stageStartedAt?: string;
    completedAt?: string;
    uploadedBytes: number;
    uploadTotalBytes?: number;
    attempt: number;
    maxAttempts: number;
    nextRetryAt?: string;
    fallbackActive: boolean;
    receiptId?: number;
    errorCode: string;
    errorMessage: string;
    canUpload: boolean;
    canRetry: boolean;
}
export namespace RecognitionTask {
}